requirements.txt
click==8.1.3
Flask==2.2.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.1
zipp==3.8.1
Flask-SQLAlchemy==3.0.2
gevent==22.10.2
requests==2.28.2